+++ src/test/groovy/ru/vyarus/java/generics/resolver/SpecialMapsTest.groovy
@@ -4,15 +4,31 @@ import ru.vyarus.java.generics.resolver.context.container.ParameterizedTypeImpl
 import ru.vyarus.java.generics.resolver.error.UnknownGenericException
 import ru.vyarus.java.generics.resolver.support.Base2
 import ru.vyarus.java.generics.resolver.util.TypeToStringUtils
+import ru.vyarus.java.generics.resolver.util.map.EmptyGenericsMap
 import ru.vyarus.java.generics.resolver.util.map.IgnoreGenericsMap
 import ru.vyarus.java.generics.resolver.util.map.PrintableGenericsMap
 import spock.lang.Specification
+
+import java.lang.reflect.Type
 
 /**
  * @author Vyacheslav Rusakov
  * @since 13.05.2018
  */
 class SpecialMapsTest extends Specification {
+
+    def "Check empty map behaviour"() {
+
+        when: "adding to empty map"
+        EmptyGenericsMap.getInstance().put("d", null)
+        then:
+        thrown(UnsupportedOperationException)
+
+        when: "adding to empty map 2"
+        EmptyGenericsMap.getInstance().putAll(new HashMap<String, Type>())
+        then:
+        thrown(UnsupportedOperationException)
+    }
 
     def "Check ignoring map behaviour"() {
 
@@ -27,6 +43,16 @@ class SpecialMapsTest extends Specification {
                 new IgnoreGenericsMap(["K": String]))
         then:
         res == "Base2<String, Object>"
+
+        when: "adding to map"
+        IgnoreGenericsMap.getInstance().put("d", null)
+        then:
+        thrown(UnsupportedOperationException)
+
+        when: "adding to map 2"
+        IgnoreGenericsMap.getInstance().putAll(new HashMap<String, Type>())
+        then:
+        thrown(UnsupportedOperationException)
     }
 
     def "Check printable generic names"() {